feat(ewm-main): event

+++ ewm-main-service/src/main/java/ru/practicum/ewm/event/model/Event.java
@@ -42,5 +42,5 @@ public class Event {
     @Enumerated(EnumType.STRING)
     private EventState state;
     private String title;
-    private int views;
+    private Long views;
 }
